backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1

# File Storage
Pillow==10.2.0
django-storages==1.14.2
boto3==1.34.34

# QR Code & PDF Generation
qrcode==7.4.2
reportlab==4.0.9
pypdf==4.0.1

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-decouple==3.8
django-filter==23.5
celery==5.3.6
redis==5.0.1

# Notifications
twilio==8.12.0

# Development
black==24.1.1
flake8==7.0.0
ipython==8.21.0
django-debug-toolbar==4.3.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.6.0
